Created configuration package and moved classes around
