+++ storage/api/impl/DataFrameSamplesQuery.py
@@ -5,7 +5,7 @@ import pandas as pd
 from storage.api.SamplesQuery import SamplesQuery
 from storage.api.impl.QueriesCollection import QueriesCollection
 from storage.api.impl.SamplesQueryIndex import SamplesQueryIndex
-from storage.connector.DataIndexConnection import DataIndexConnection
+from storage.configuration.connector import DataIndexConnection
 from storage.variant.SampleSet import SampleSet
