+++ storage/test/integration/configuration/connector/test_DataIndexConnection.py
@@ -4,7 +4,7 @@ from tempfile import TemporaryDirectory
 
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 
-from storage.connector.DataIndexConnection import DataIndexConnection
+from storage.configuration.connector.DataIndexConnection import DataIndexConnection
 from storage.test.integration import sample_dirs, reference_file
 from storage.variant.io.mutation.NucleotideSampleDataPackage import NucleotideSampleDataPackage
 from storage.variant.io.processor.SerialSampleFilesProcessor import SerialSampleFilesProcessor
